Update _search.html.erb with suggestions.
Usually, form_for is used to create or update a specific Model. In this case, we're just wanting to create a form that we will send using AJAX. So a better form helper to use here would be form_tag. Also, your "style" attribute on the submit_tag should go in an external css file. :)